pyproject.toml
[tool.poetry]
name = "mini-app-api"
version = "0.1.0"
description = ""
authors = ["zytfo <khayalievartur@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.21.1"
alembic = "^1.10.2"
fastapi = "^0.111.0"
passlib = "^1.7.4"
pyjwt = "^2.6.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.22"
python-multipart = "^0.0.9"
asyncpg = "^0.28.0"
greenlet = "^3.0.1"
pre-commit = "^3.1.1"
aiohttp = "^3.8.4"
aiosmtplib = "^2.0.2"
pip = "^23.0.1"
install = "^1.3.5"
pydantic = {extras = ["email"], version = "^2.4.2"}
aiofiles = "^23.1.0"
psycopg2-binary = "^2.9.6"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
celery = {extras = ["redis"], version = "^5.3.1"}
prometheus-client = "^0.17.1"
opentelemetry-instrumentation-fastapi = "^0.40b0"
opentelemetry-instrumentation-logging = "^0.40b0"
opentelemetry-exporter-otlp = "^1.19.0"
opentelemetry-api = "^1.19.0"
opentelemetry-sdk = "^1.19.0"
httpx = "^0.24.1"
pillow = "^10.0.1"
pytest = "^7.4.2"
aiohttp-retry = "^2.8.3"
bump-pydantic = "^0.7.0"
slowapi = "^0.1.8"
redis = ">=4.2.0rc1"
trio = "^0.23.1"
jsondiff = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"